Last production bugs.
Fixed input validation tests on docker-commit.sh. Implemented test procedure before doing mysql or psql dumps. Implemented rotation for sql dumps.

+++ docker-commit-dump.sh
@@ -11,19 +11,20 @@ ORGANIZATION=$2
 IMAGE_NAME=$3
 BKPATH=$4
 
-for var in {CONTAINER_ID,ORGANIZATION,IMAGE_NAME,BKPATH}
+for var in {"CONTAINER_ID","ORGANIZATION","IMAGE_NAME","BKPATH"}
 do
-    if [ -z $$var ]
+    eval testvar=\$$var
+    if [ -z $testvar ]
     then
 	echo No $var provided.
-	return 1
+	exit 1
     fi
 done
 
 if [ ! -w $BKPATH ]
 then
     echo $BKPATH does not exist or is not writable.
-    return 1
+    exit 1
 fi
 
 #Get timestamp in seconds from epoch format
@@ -31,12 +32,28 @@ DATE=`date +%s`
 
 #Commit, dump and gzip routine
 
-echo Commiting container $CONTAINER_NAME
+echo Commiting container $CONTAINER_ID to image $IMAGE_NAME
 docker commit -m="Backing up container $CONTAINER_NAME" -a="Cron Backup Job" $CONTAINER_ID $ORGANIZATION/cron-backup-job-$IMAGE_NAME:$DATE
-echo Saving container $CONTAINER_NAME
+echo Saving container $CONTAINER_ID to image $IMAGE_NAME
 docker save $ORGANIZATION/cron-backup-job-$IMAGE_NAME:$DATE | gzip > $BKPATH/cron-backup-job-$IMAGE_NAME-$DATE.tar.gz
-echo Dumping all psql on container $CONTAINER_NAME
-docker exec $CONTAINER_ID pg_dumpall --username=postgres | gzip > $BKPATH/cron-backup-job-$IMAGE_NAME-$DATE.psql.gz
+
+#Test for PostgreSQL and dump it (service has to accept passwordless access)
+echo Testing for psql
+docker exec $CONTAINER_ID psql -Upostgres -l 1>/dev/null
+if [ $? -eq 0 ]
+then
+    echo Dumping all psql on container $CONTAINER_ID to dump $IMAGE_NAME
+    docker exec $CONTAINER_ID pg_dumpall --username=postgres | gzip > $BKPATH/cron-backup-job-$IMAGE_NAME-$DATE.psql.gz
+fi
+
+#Test for MySQL and dump it (service has to accept passwordless access)
+echo Testing for mysql
+docker exec $CONTAINER_ID service mysql status 1>/dev/null
+if [ $? -eq 0 ]
+then
+    echo Dumping all mysql on container $CONTAINER_ID to dump $IMAGE_NAME
+    docker exec $CONTAINER_ID mysqldump --all-databases | gzip > $BKPATH/cron-backup-job-$IMAGE_NAME-$DATE.mysql.gz
+fi
 
 #Rotate commits, saves and dumps
 #Keeps the current and the last
@@ -52,7 +69,18 @@ do
     rm -v $IMTAR
 done
 
-for DPSQL in `ls -t $BKPATH/cron-backup-job-$IMAGE_NAME-*.psql.gz | tail -n +3`
-do
-    rm -v $DPSQL
-done
+if ls $BKPATH/cron-backup-job-$IMAGE_NAME-*.psql.gz 1>/dev/null 2>&1
+then
+    for DPSQL in `ls -t $BKPATH/cron-backup-job-$IMAGE_NAME-*.psql.gz | tail -n +3`
+    do
+	rm -v $DPSQL
+    done
+fi
+
+if ls $BKPATH/cron-backup-job-$IMAGE_NAME-*.mysql.gz 1>/dev/null 2>&1
+then
+    for MYSQL in `ls -t $BKPATH/cron-backup-job-$IMAGE_NAME-*.mysql.gz | tail -n +3`
+    do
+	rm -v $MYSQL
+    done
+fi
